File Share : api, service layer implementations

+++ server/src/main/java/org/apache/cloudstack/storage/fileshare/FileShareServiceImpl.java
@@ -25,16 +25,8 @@ import java.util.Map;
 import javax.inject.Inject;
 import javax.naming.ConfigurationException;
 
-import com.cloud.exception.ResourceUnavailableException;
-import com.cloud.storage.VolumeApiService;
-import com.cloud.storage.dao.VolumeDao;
 import com.cloud.utils.Pair;
-import com.cloud.utils.db.Filter;
-import com.cloud.utils.db.SearchBuilder;
-import com.cloud.utils.db.SearchCriteria;
-import com.cloud.vm.VirtualMachine;
-import com.cloud.vm.VirtualMachineManager;
-import com.cloud.vm.dao.VMInstanceDao;
+import com.cloud.utils.db.DB;
 import org.apache.cloudstack.api.command.user.storage.fileshare.CreateFileShareCmd;
 import org.apache.cloudstack.api.command.user.storage.fileshare.ListFileShareProvidersCmd;
 import org.apache.cloudstack.api.command.user.storage.fileshare.ListFileSharesCmd;
@@ -47,9 +39,7 @@ import com.cloud.event.EventTypes;
 import com.cloud.user.Account;
 import com.cloud.user.AccountManager;
 import com.cloud.utils.component.ManagerBase;
-import com.cloud.utils.db.DB;
 import com.cloud.utils.exception.CloudRuntimeException;
-import org.apache.cloudstack.storage.fileshare.dao.StorageFsVmDao;
 
 public class FileShareServiceImpl extends ManagerBase implements FileShareService {
 
@@ -57,22 +47,7 @@ public class FileShareServiceImpl extends ManagerBase implements FileShareServic
     private AccountManager accountMgr;
 
     @Inject
-    VirtualMachineManager virtualMachineManager;
-
-    @Inject
-    private VolumeApiService volumeApiService;
-
-    @Inject
     private FileShareDao fileShareDao;
-
-    @Inject
-    StorageFsVmDao storageFsVmDao;
-
-    @Inject
-    VMInstanceDao vmDao;
-
-    @Inject
-    private VolumeDao volumeDao;
 
     protected List<FileShareProvider> fileShareProviders;
 
@@ -142,11 +117,11 @@ public class FileShareServiceImpl extends ManagerBase implements FileShareServic
     @Override
     @DB
     @ActionEvent(eventType = EventTypes.EVENT_FILESHARE_CREATE, eventDescription = "Deploying fileshare", create = true)
-    public FileShare deployFileShare(Long fileShareId, List<Long> networkIds) {
+    public FileShare deployFileShare(Long fileShareId, Long networkId) {
         FileShareVO fileShare = fileShareDao.findById(fileShareId);
         FileShareProvider provider = getFileShareProvider(fileShare.getFsProviderName());
         FileShareLifeCycle lifeCycle = provider.getFileShareLifeCycle();
-        Long vmId = lifeCycle.deployFileShare(fileShare, networkIds);
+        Long vmId = lifeCycle.deployFileShare(fileShare, networkId);
         fileShare.setVmId(vmId);
         fileShareDao.update(fileShare.getId(), fileShare);
         return fileShare;
@@ -164,40 +139,38 @@ public class FileShareServiceImpl extends ManagerBase implements FileShareServic
 
     @Override
     public Pair<List<? extends FileShare>, Integer> searchForFileShares(ListFileSharesCmd cmd) {
-        final SearchBuilder<FileShareVO> sb = fileShareDao.createSearchBuilder();
-        sb.and("id", sb.entity().getId(), SearchCriteria.Op.EQ);
-        sb.and("account_id", sb.entity().getAccountId(), SearchCriteria.Op.EQ);
-
-        Filter searchFilter = new Filter(FileShareVO.class, "id", Boolean.TRUE, cmd.getStartIndex(), cmd.getPageSizeVal());
-        final SearchCriteria<FileShareVO> sc = sb.create();
-
-        Long id = cmd.getId();
-        Long accountId = cmd.getAccountId();
-        if (id != null) {
-            sc.setParameters("id", id);
-        }
-        if (accountId != null) {
-            sc.setParameters("account_id", accountId);
-        }
-
-        Pair<List<FileShareVO>, Integer> result = fileShareDao.searchAndCount(sc, searchFilter);
+        Pair<List<FileShareVO>, Integer> result = fileShareDao.searchAndCount(cmd.getId(), cmd.getAccountId(), cmd.getNetworkId(), cmd.getStartIndex(), cmd.getPageSizeVal());
         return new Pair<List<? extends FileShare>, Integer>(result.first(), result.second());
+    }
+
+    @Override
+    @DB
+    @ActionEvent(eventType = EventTypes.EVENT_FILESHARE_CREATE, eventDescription = "Updating fileshare", create = true)
+    public FileShare updateFileShare(UpdateFileShareCmd cmd) {
+        Long id = cmd.getId();
+        String name = cmd.getName();
+        String description = cmd.getDescription();
+
+        FileShareVO fileShare = fileShareDao.findById(id);
+        if (name != null) {
+            fileShare.setName(name);
+        }
+        if (name != null) {
+            fileShare.setDescription(description);
+        }
+        fileShareDao.update(fileShare.getId(), fileShare);
+        return fileShare;
     }
 
     @Override
     public FileShare deleteFileShare(Long fileShareId, Account owner) {
         FileShareVO fileShare = fileShareDao.findById(fileShareId);
 
-        Long vmId = fileShare.getVmId();
-        if (vmId != null) {
-            VirtualMachine vm = vmDao.findById(vmId);
-            try {
-                virtualMachineManager.expunge(vm.getUuid());
-            } catch (ResourceUnavailableException e) {
-                logger.warn(String.format("Unable to destroy storagefsvm [%s] due to [%s].", vm, e.getMessage()), e);
-                return null;
-            }
-            storageFsVmDao.remove(vmId);
+        FileShareProvider provider = getFileShareProvider(fileShare.getFsProviderName());
+        FileShareLifeCycle lifeCycle = provider.getFileShareLifeCycle();
+        boolean result = lifeCycle.deleteFileShare(fileShare);
+        if (!result) {
+            return null;
         }
         fileShareDao.remove(fileShareId);
         return fileShare;
